Add files via upload
This is a simple Python script that sends notifications every hour, reminding the user to take a break. The script uses the plyer module to display notifications on the operating system. When executed, the script enters an infinite loop, sending a notification every hour with the title "Hey, you." and the message "You're finally awake". After displaying the notification, the script waits another hour before sending the next notification. This script can be useful for maintaining a healthy work routine, reminding the user to take regular breaks throughout the day.